cleanup secondary sidebar code
