+++ app/src/componentsV2/SecondarySidebar/SecondarySidebar.css
@@ -1,8 +1,8 @@
-.rules-sidebar-container {
+.rq-secondary-sidebar-container {
   height: 100%;
   background-color: var(--component-background);
 }
 
-.rules-sidebar-container ul {
+.rq-secondary-sidebar-container ul {
   padding: 10px 10px 1.25rem 10px;
 }
