+++ app/src/componentsV2/SecondarySidebar/index.ts
@@ -1,0 +1,2 @@
+export { default as SecondarySidebar } from "./SecondarySidebar";
+export * from "./components/SecondarySidebarItem/SecondarySidebarItem";
